Friends

+++ GAMMA/src/main/java/org/game/user/domain/QuestionVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class QuestionVO {
-	private Long qnum;
+	private int qnum;
 	// 계정 및 내정보, 소프트웨어 및 게임, 라이브러리, 구매, 친구, 채팅, 어플리케이션, 기타
 	private String qtype;
 	private String qtitle;
